Fix run query from gutter icon in Kotlin file.
Usign PsiElement.copy works fine for instances of RoomSql in
Java files, but does not work for instance of RoomSql in Kotlin
files.
Cloning the view provider of the RoomSql PsiElement returns null.
This happens only in the IDE, everything works fine in unit tests,
therefore I have not been able to reproduce the issue in tests.

Test: can't reproduce in tests
Bug: 152116515
Change-Id: I3aa06c234b9221d2e4cbe36c12a3754e7b22aa5b