Address dispose race condition
When obtaining the rendered image, if the RenderResult is disposed
before the image copy is obtained, this would fail, trying to obtain a
copy of the image.
If the RenderResult is disposd, the returned pool image should be
NULL_POOLED_IMAGE.
After this, we ensure that in render result is disposed we return the
null image.

Fixes: 151807086
Test: N/A. There is no easy way to manully reproduce the race condition.
Change-Id: Ifde5fbf0a7d2747965ec6740ab64f73764f8c096
(cherry picked from commit 28ce159c674f2890841af2c8d25d999e78aa8efd)