Simplify AtraceParser constructors
This keeps one very simple for test that doesn't care about which
process while we're selecting, while moving the other ones to the
main constructor.

We also added default values to MainProcessSelector in order to make it
easy to represent (in Kotlin) when you want to select only based in one
criteria.

Bug: n/a
Test: existing
Change-Id: Ida32bccf3f7ec182a7524258102bb8b558958256