[Dsl,.kts] fix extra prefix interaction with apply
The logic for detecting when an extra property reference in
KotlinScript needs a prefix was only correct for projects with one
build file per module.  However, applied files break this logic, and
instead of a simple non-equality test, we need to test whether the
main build file at any point transitively applies the file defining
the property being referred to, and only if this is not the case
should we consider parent projects.

Bug: 142114586
Bug: no specific bug logged
Test: added
Change-Id: Iea757b6b8e62e0e8904811a3800146a1d4df9582