Make ProfilingConfiguration fields NotNull
All users of .getName(), .getTraceType() and .getMode() were already
assuming the returned values were not null.

We only use the default constructor in one place, in CpuCaptureParser
when the configuration is missing as a fallback, usually because it's an
imported trace, so we don't have a configuration for it.

This prevents a bug, when we try to report the stats for such cases and
we end up throwing a NPE because the switch statement in
StudioFeatureTracker doesn't handle null values.

Bug: n/a
Test: existing + manually (because FakeIdeProfilerServices masks the bug)
Change-Id: I6606bfdcf12ac20300b0e2a396e671779028f6e6