Fix incorrect cancellation of AdbService internals
When the connection to the ADB server is disconnected, systems
will reattempt to get the bridge again from AdbService. Within
AdbSerice#getBridge, AdbService will detect that the bridge is
stale and attempt to destroy it prior to recreating the bridge,
on a different thread.

However, deep down in the destruction code, there is a call
that invalidates a future associated with the ongoing session.
Normally, this termination behavior is correct with respect to
external callers. However, for getBridge, this actually
invalidates the newly created future instance for the new
session, not the stale future from the previously disconnected
session.

This fix just moves the cancellation to the external version of
the method (the one that outside code actually calls) and
renames the "internal" version.

This change was rebased and merged without the following changes:
* Icbcd6bbdb7edfb2b3741048e8903b4afe2cc332a Add timeout when starting
  and terminating ADB

Fixes: 152351487
Test: Added test for reinitializing the bridge.
Change-Id: Ib844f0301a71ae854c657afec4779f8d261a6e99
(cherry picked from commit 6a6d67ae4f9784b091735789570432c4103ec109)