[Dsl,.gr] fix resolving references with exotic names
Define a gradleNameFor() utility function, by loose analogy with the
KotlinScript utility of the same name.  The Groovy implementation can
be simpler, given the correspondence between the Groovy syntax and the
GradleNameElement syntax; the work that needs to be done is
identifying the precise name corresponding to a given Groovy reference
expression, handling any unquoting necessary, and then escaping the
result appropriately for GradleNameElement syntax.

This fixes at least one bug in the PSD, where references to e.g. a
signingConfig with a non-alphanumeric name would show up in red as
failing to resolve.

Bug: 135224836
Bug: 152603582
Test: add gradleNameFor unit tests for Groovy
Test: manual test of PSD resolution
Change-Id: Ia7ec82b06193e7e31e245ffd52b0c01b044281e4