Workaround test terminated error in debug run
AndroidRemoteDebugProcessHandler's lifecycle is tied with
a Debug run session. If you debug android instrumentation
test process, the test process may terminate before Ddmlib
calls testRunEnded callback. This confuses the IDE and it
thinks the test run is terminated before the test finishes
and shows an error message. This patch adds 1 second delay
to notifyProcessDetached() callback in DebugProcessListener
to give Ddmlib a chance to call testRunEnded() as a
workaround solution.

Test: manual testing
Bug: 150001290
Change-Id: If7fa3820b4ea80935d4aea26aff6bc31ee902ba1