Use accumulator in applyFilter calculation
The way we're using FilterResult inside CaptureNode recursive
computation for applyFilter, meant we were creating an instance
of it for each node, just to discard it when we accumulated the
values on the local variables (that then we used to create the new
ResultFilter to return in the recursion, etc ...).

We get rid of all those unecessary allocations by introducing
FilterAccumulator which only job is to accumulate those counters and
only produce a FilterResult on the last step.

Bug: 152963382
Test: existing
Change-Id: I26add0f387715b97540879059c05fb4a34ad1004