[Sync Error Handlers] Convert more ErrorHandlers.
In this CL:
Create OldAndroidPluginIssueChecker to offer quickFixes to upgrade AGP
when the version is too old.
Also, delete ObjectStreamErrorHandler because it was normally used to
provide a better message to users when they get a sync issue raised on Gradle
side ("unexpected end of block data") but no quickFixes.

Bug: 149822714
Bug: 149822714
Test: added
Change-Id: Ia0a0a8b26ab7e5b41cc7a027847484ff5cb6f310