Rework the way we wait for sync to finish in UI tests
(1) Ensure waiting is only possible for code which triggers sync
(2) Make it impossible to wait for sync which hasn't been invoked

This change changes the pattern of triggering and waiting for sync in
UI tests from:
  ```
  requestSync() // Directly on indirectly
  ...
  waitForSyncToFinish()  // Which maybe invoked several times
                         // without requesting another sync.
  ```
to
  ```
  actAndWaitForSyncToFinish {
      requestSync() // directly or indirectly
  }
  ```

The new waiting pattern will wait for sync initiated within the block
of code only.

This change do not change the mechanism of waiting yet.

Bug: n/a
Test: n/a
Change-Id: I58f670956a2f278e04cb9ba82c9c2562dc04c74b