Remove more references to SceneView in layout
This is the second step in migrating out from using directly SceneViews
and moving more code to Swing. This just refactors the the existing code
to remove direct references to SceneView when possible and to use
PositionableContent.

The next step will be to remove the PositionableContent class and move
all the functionality to the SceneViewPeerPanel.

Bug: N/A
Test: Existing tests pass
Change-Id: If5902f9b030ba844e639748d0fa596c6597f32dc