Verify project is not disposed before sync after
... when syncing after a build.

When the user request a sync while building, or one is needed after a
build, we do that by creating a sync request using
Application.invokeLater, but it is possible that some of the tasks stils
pending causes the project to be disposed, for example, if the user
decides to close the project.

This CL confirms that the project has not been disposed before starting
sync when it is needed after a build.

Bug: 153375399
Test: Manually reproduced the problem and confirm that original exception
is not thrown.

Change-Id: Id3c34da6ca404e4582fbb73aa5bcf8ad16879727