Add action for opening the device pairing dialog
This is mostly glue code, except for the "AdvDevicePairingPanel" code,
which will need some ravamping later on (cleanup swing layout).

The purpose of this change is mostly to get this glue code "out of
the way" so incoming CLs can be smaller.

This change also adds a new studio flag (ADB_WIRELESS_PAIRING_ENABLED)
to control the visibility of this feature.

We need to add 2 dependencies to the android-adb module:

* android.sdktools.flags: needed the "Flag" class definition
* android-common: needed to access "StudioFlags"

The big change to the BUILD file is due to adding a dependency on
"android-common"

Bug: 150796011
Test: Included
Change-Id: I83b76de71f17a67e85ef21591018a6d9c62eb2ee