Tweak timeline to indicate no live info when taking heap dump
This CL adds the ability to register a custom renderer to
`DurationDataRenderer` to do arbitrary drawing on top of
the standard drawing.

The `heapDumpRenderer` then uses that to draw a blank box
(white or dark gray, depending on theme) with 2 vertical bars
to indicate the lack of live information when taking the
heap dump.

Bug: 153690080
Test: add new
Change-Id: I6951bef23e8b30664ebe874aa22d14671113a00e