Avoid refreshing combobox twice
When a new process is added to an empty combobox, we unwittingly fire ItemEventListener twice. Once when the selection changes from "No Process Available" to "No Process Selected", and once more when selection changes from "No Process Selected" to "Process X".

So now we check to see if the item is a ProcessDescriptor before we refresh the combobox. This CL breaks down the existing refreshTabs function into two distinct functions, one for clearing existing inspector tabs, and one for populating new inspector tabs.

Test: no functional change. Tested manually.
Bug: N/A
Change-Id: I585ce1d0ba54aaf54702c3cb627603cfa248b546