Lazy split editor preview initialization
Now, we will only initialize the preview panel if it is open. If the
panel is not visible, the initialization will be deferred until later.
If the user has selected that they want Code mode as default, then the
initialization will not happen until they change the tab.

This CL moves the split editor state from the DesignSurface to the
DesignerEditorPanel. The DesignSurface was just acting as a holder and
the only elements interested were analytics and DesignEditorPanel.
Moving the state to the DesignerEditorPanel allows to remove the storage
from the surface.

Fixes: 154091487
Test: Updated tests pass.
Change-Id: I00ea8468a28848b8956683bc7deb6684d1c17ed4