Disable checks for file access outside roots in Bazel
Tests are running in a sandbox, and there's no point in checking file
accesses. There is a difference in performance, however, if we
specify overly-broad roots and still let the checks happen, especially
on Windows where I/O is expensive.

With this change we disable the checks just like when running from the
IDE, via NO_FS_ROOTS_ACCESS_CHECK.

Test: N/A
Bug: N/A
Change-Id: I68d21473c950946d23cfa4f82567b96f42592361