Build UI panel to display Android Test Retention snapshots
This CL builds a tab in test results to display Android Test Retention
snapshot artifacts. The tab will only show up if there is a retention
artifact imported from the UTP test result protobuf. Thus it has
implicit dependency on the Android Studio flag
utp.instrumentation.tests.

Currently, the panel only displays a huge button in the middle, and is
supposed to be changed once we have the UI mock.

Bug: 152773288
Test: AndroidTestResultsTableViewTest, DetailsViewContentViewTest, UtpTestResultAdapterTest
Change-Id: I79417291c1bd8dcf0095460b095402e4399e8e56