Mitigate errors in closed stream channel
We have received reports of IllegalStateException throws in TransportStreamManager because the service is trying to dispose inspectors right after the process they are attached to ended. When a process ends, it's too late to send commands and listen for response. In that circumstance, we want to avoid sending commands as much as possible

This CL makes it so the service actively cleans up clients associated with terminated process, before it notifies interested parties of said termination. Therefore, we can guard against calls to dispose already-terminated clients by making them no-ops.

With these measures in place, we can still hit this exception due to pure bad luck in timing. Therefore, I changed the exception to a logged error.

Test: new AppInspectionDiscovey test.
Bug: 154460180
Change-Id: Ife87411bf2cba3ce41daf54cd078e22936accf78