Fixed issue with light classes and base classes
It seems like logic was somewhat recently changed in IJ code to start
calling PsiClass.getSupers(). This was previously not handled by the
data binding light class implementation.

This CL fixes that as well as tightens up a few other APIs related to
base class resolution, to avoid future surprises.

Fixes: 154652005
Test: Tests re-enabled, LightBindingClassTest
Change-Id: I4a926565f77a41369dc6b1adac1b76a648b66c8c