Fix flaky ComposeDocumentationProviderTest
The ComposeDocumentationProvider implements DocumentationProviderEx that
requires returning results immediately (since it runs on the UI thread).
During testing, the image might not be rendered on time and the method
would return null, which is expected.
During tests, we want to wait for the image to arrive. This CL exposes
the CompletableFuture to allow for that.

Fixes: 155259494
Test: Executed ComposeDocumentationProviderTest 70 times. It used to
fail 1 in 10

Change-Id: I465e1ee1f302611f12ea4aeba45a20ff007e782f