[TP] Introduce SystemTraceModelAdapter
SystemTraceModelAdapter is a data adapter that provides a common
interface to system trace data, independent from which backend it's
coming from (Trebuchet or Perfetto).

This way we can abstract all the computation done inside AtraceParser
away from Trebuchet-specific data models, so we can then extract them
and use in both AtraceCapture and PerfettoCapture.

About tests: This CL is mainly about introducing the adapter and
migrating all the usages. Most of it is very simple (interface and data
classes), with the exception of TrebuchetModelAdapter which logic is
very well covered by all existing tests, which were adapted to use the
adapter instead of the Trebuchet model directly.

Bug: 147099951
Test: existing (see above)
Change-Id: I9b00775a78e4f6d6403f406634ad478deaac9c7e