use shortened name in RenameModule
We pre-populate the module name in the dialog from the short name (as
visible in the Project view UI); that shortened-style name is then
used as the Gradle project name for the new module.  Use that
shortened name both as the name for the action, simplifying entries
relating to the action in e.g. Undo menus; and in input validation,
disabling the OK button when the new name is equal to the existing
one.

Bug: 150933937
Test: existing
Change-Id: I010a0236eefb9eef315fc5ed3f5dd48d5bcf4288