Extract SystemTrace data computation to SystemTraceCpuCaptureBuilder
This CL adds SystemTraceCpuCaptureBuilder and moves all the SystemTrace
data computation from ATraceParser into it. Now that class is
responsible for computing all the data series and associate objects
(e.g. FrameManager, SurfaceflingerManager) in order to build and return
the final (SystemTrace)CpuCapture.

This allows the code to be re-used inside TraceProcessor code path.

Bug: 147099951
Test: existing
Change-Id: Iac1f36603947842887e9e96a13af9f0baca55b93