[lint] check non-single-argument Gradle method calls
If we see a single-argument method call within a closure, we treat it
as being an assignment of a Dsl property -- this is an approximation,
but for the purposes of Lint it is adequate.  However, there are
things we want to consider that are not assignments and not
single-argument method calls; if we encounter one of those, check it
as a method call.

Bug: 144442233
Test: in forthcoming CL
Change-Id: I6694a0008bc5b4ad12753f6babdad8121d4754b0