Android: fix test com.android.tools.idea.naveditor.structure.DestinationListTest.testRightClickActivity
It is not valid to expect that `NavActionManager#getPopupMenuActions` will be called on the mocked object, because this method is only called from `NavActionManager#showPopup` (i.e. the mock should call the method itself). Test was working only because the mock was partial (because `showPopup` is final method). With InlineByteBuddy mock creator the mock is not partial anymore, and the test fails.

GitOrigin-RevId: d99cf0fa225e5397e512c157fa95746532c0aac8