Remove deploy compose preview minimum version banner
Now that the newest Compose version is dev12 and Android Studio templates
set the version to dev09, it's unrealistic to think users will be depending
on dev03 or older compose versions. Therefore, we can safely remove the
banner that informs users that Deploying compose previews is only available
from dev04.

Fixes: 150391864
Test: N/A. Visual change.
Change-Id: Ic37a19acb10f1ebadd07d1ae91d860020c109ba2