Add comments and clarify indexes for dev12
With dev12 comming, most of the source line information is not from
inline methods. That means that we had a combination of 0-indexed and
1-indexed information.
Most of the code in the Compose preview should work with 1-indexed since
it eases debugging. This CL clarifies the indexes and makes sure there
are clear boundaries for the 0-indexes values.

Bug: 156589330
Test: The existing navigation tests pass
Change-Id: I7f31efe39f10ad5542855df8fa3cdcadbbf07020