[BA] Swap time value and percentage places
On the latest sync discussion it was decided that
The first argument for it is that time is probably more important and
the second is that it is easier to align numbers vertically this way
since percentage part can have a fixed width.
Also it was decided that we remove a space between the time value and
'seconds' unit to improve visual grouping of these two parts.

Bug: 155871850
Test: Updated
Change-Id: I21122eeeefd727f3d900f39653762ddc05e2517e