Use '\n' instead of platform dependent line separator in sample code
It's unnecessary to be platform specific in the sample code case
because EditorTextField will covert "\r\n" if any to "\n" in its
constructor anyway. Also on Windows, the conversion somehow doesn't
work as expected and causes b/156460170. So I think we can just
always use '\n' in first place to make things eaiser.

Bug: 156460170
Test: No Windows device on hand, but have contacted with QA to verify
the change once it is available in the master build
Change-Id: I31cb3cbb2d79ffbf93df44c90683bc5d6381502d