Fix the Color gutter Icon for Compose
This is a quick fix to adjust to the changes of the 'Color' class.

The Color class is now a 'ULong' inline Class, so from the Annotator
perspective, the helper methods used to create Color instances now
return a 'long' value.

To verify that the methods are the correct ones to be supported, we now
check that it's part of the ColorKt.class

Bug: 156796445
Test: ComposeColorAnnotatorTest.kt
Change-Id: Ie73c1a80d38716efdf3519aea18521ab54d875cc