Deprecate TransactionGuard.submitTransaction usage
All the changes here are related to the deprecation of the
TransactionGuard.submitTransaction() methods related to this cherry
picked commit from JetBrains: http://go/androidcommit/53ee762 as part of
the 2020.1 merge.

For detailed information on how to adapt to the deprecated API, and an
explanation of the IDEA threading model (write-safe contexts, locks,
etc) please read this document: http://go/idea-threading-model

1. setSchemaFromSelectedItem() runs on the EDT, and is already in a
write-safe context (since the code runs as a result of user activity in
a non-modal state). Also dropPsiCaches() itself uses a write action to
do its job, and there's no need to wrap it in a write action.

The following tests pass:

- SqliteEvaluatorViewImplTest.kt

Bug: 156102335

Change-Id: I5df01bbe2b3963a3200aaf61f81b5f8b9bca764e