Fix heap dump rendering and its tooltips on timeline
This CL removes the unneccesary drawing for the heap dump from
inheriting `DurationDataRenderer` (now that it's just a blank box
with no line), and stop drawing the "memory usage" tooltip and
the veritcal bar when the mouse hovers over any of the heap dump.

More specifically, it does the following:

* Declare an interface `AbstractDurationDataRenderer` that extends
`LineChartRenderer`, and rebase `DurationDataRenderer` to implement
this new interface

* Remove the previous unwieldy mechanism of `customDecorator` in
`DurationDataRenderer`

* Make a separate subclass `HeapDumpRenderer` that implements
`AbstractDurationDataRenderer` that:
  - handles the drawing (previously in `HeapDumpDataRendering.kt`
  - records when the mouse hovers over any of the heap dump

* Allow signaling with `DurationDataRenderer` that the range is
being occluded, so that it switches to the default cursor

This CL also fixes b/156790753 by accident.

Fixes: 154171109
Fixes: 154264530
Fixes: 156790146
Fixes: 156548001
Fixes: 156790753
Test: add new, remove obsolete
Change-Id: I3f934792355ed46ae4a0dffdd77e90fe2d34d9aa