Guard against null if the target directory is null in Fragment wizard
The virtual file becomes null if the destination in navigation editor
is deleted using backspace as explained in b/156771344.
Eventually that led to a crash in the Fragment wizard invoked from
navigation editor.

The root cause should be fixed by b/156771344, but this CL guard
against null even if the virtual file becomes null for now.

Bug: 154588600
Test: Deleted a destination in nav editor using backspace, then
      select "Create new destination" and verified it doesn't crash
Change-Id: I62ed69928e7a0b994a5e93556a1221f46711cb8e