[Windows] Fix AVD Manager missing header
AVD Manager is the only window specific to Studio that extends
FrameWrapper. The title bar and header were missing because the
getComponent method in AvdListDialog was overriding both parents: one in
FrameWrapper and one in AvdUiAction.AvdInfoProvider. Overriding the
former was causing the title bar to disappear, because the expected
component there should have been the root pane, which is already being
set correctly in the init. So this CL renames the getComponent in
AvdInfoProvider to avoid the name clash.

Bug: 155905906
Test: manual
Change-Id: Ic2267e260edf996951fd94690b67e9cbc7dbc312