Add unique ID to FakeGrpcChannel name
In a test we are writing, we found that when multiple tests run at the
same time using a FakeGrpcRule with the same name, sometimes messages
fired in one test are picked up in another. In order to prevent
different tests from affecting each other due to accidentally "shared"
state, we append a UUID to the end of each GRPC connection.

Bug: 157174381
Test: Only test code updated
Change-Id: I8206fc3b5262f465d13a763399fe010ec934c4de