Remove deactivation listener from NlModel
The activation/deactivation listening should not be done at the model
level since it's a property of the surface (is it visible or not?). This
allows removing methods from the already over used ModelListener and to
concentrate the listening closer to the responsability. The model will
still get the notification from the SceneManager when
activation/deactivation happens so it can subscribe/unsubscribe to
events.

The next step is to remove the activation listening since it's only used
to refresh the UI when the panels come into view.

Bug: N/A
Test: Existing tests pass
Change-Id: I32a1faa33af8da2485a6e73e21829852326c132d