[AUA] renames, comment fixes
Rename AgpUpgradeVersionRefactoringProcessor to
AgpUpgradeRefactoringProcessor, to reflect the refactoring being
related to an AGP Upgrade, not just a version upgrade.

Rename AgpUpgradeUsageInfo to GradleBuildModelUsageInfo, to reflect
the participation of generalized instances in the GradleBuildModel
protocol: their contract in their methods is to manipulate
the (shared) build model so that the
GradleBuildModelRefactoringProcessor can do the right thing.

Bug: 149565531
Test: existing
Change-Id: I3788fda28af42a5078c11b4a5442e2d5ea634bdb