Render primary key icon in table view and not other icons
This CL removes column icons from the headers of the table,
except for primary key icon, that we want to render.

A new class ViewColumn is added, that contains view information
for a column. In the controller ResultSetSqliteColumns
are mapped to ViewColumns.

ResultSetSqliteColumn might contain parial information, becuase
the live inspector is not able to provide some data
in result sets.
To compensate for this limitation this CL tries to
get the missing data from the schema.

Test: TableControllerTest.kt
Bug: 156911745
Change-Id: Id69a60e92a04edf8fc9a1dcb4d9589b66dc91c0f