Prevent updating db is cell is edited with same value
If a table cell is edited with the same value we should not
run an update on the db.

This is common when the user double clicks on a cell just
to copy its value.

Test: TableViewImplTest.kt
Bug: 157093008
Change-Id: Ied71ac9b3584e9337530396e84a6b16f228e5f0e
(cherry picked from commit 944dbe9b57a779667a9d0a05fd4195cf8e69b4d0)