fixing CoordinatorLayoutHandlerTest tests
After the recent change in the way Task.Backgroundable is executed in the tests mode, we've made the behaviour closer to the production and the task is no longer executed with a higher modality state (as if a modal progress is shown). That change may break some tests where it was assumed no invokeLater is possible between several steps.

In the the current test we've started to see exceptions in the catch block catching the PsiInvalidElementAccessException in PsiResourceItem constructor leading to not fully initialized object. (e.g. the scheduled task itself my outlive project or test).

The original cause of the tests was in the com.android.tools.idea.res.PsiResourceItem.getConfiguration assert after an attempt to compute the getKey() for the object to be removed.

The sourceFile is set in the addItem method of the PsiResourceFile class. The set sourceFile is effectively used to compute the getKey() (via getConfiguration() and getSourceFile() methods).
\Setting the sourceFile to null (in the removeItem method of PsiResourceFile) forces the class to re-compute the key (possible differently) and that code path is failing (also in the tests) on the PsiResourceItem objects for which PsiInvalidElementAccessException was thrown.
Flipping the two statements makes the tests work correctly and probably fixes a production issues too

(cherry picked from commit cf0de2b9af2e3b1e1df89cd500259be7b7a41b50)

Change-Id: I29a6196b859463598bbc928c8759570beb0ff5ad