Rename service for clarity and make it public
Currently, we only support launching inspectors indirectly through the
AppInspectionView, but we also want to support the ability to launch an
inspector even using your own custom UI (e.g. LayoutInspector might use
this). Having the ability to do this also simplifies a tutorial I'm
writing, which lets the first part focus just on the inspector and not
worry about the UI framework.

Bug: N/A
Test: Covered by existing
Change-Id: I699c1c75f91e819cb486f88e8720b9ab872e60c2