[Gradle Build] Fix ensuring buildSrc folders are ignored
These are part of the main build, we do not need to run tasks for the
separately. This was broken with the move to qualified names as the comparison was base on the full name of the intellij module.

Bug: 158018185
Test: unit test added
Change-Id: I3e1e5d78f2b7d9ba0c56d8977a04252755a93022
(cherry picked from commit d274fc21f261c4a758e735446c23d5ec6bc7d1cc)