[Dsl,.kts] renaming named domain objects is tricky
If we rename a named domain object to or from one of the names that is
implicitly created by a plugin, the way it is configured in
KotlinScript changes between create(..) and getByName(..).  When
updating such a name in the KotlinScript writer, check to see whether
its method needs updating, and if so perform the necessary Psi
gymnastics to ensure that it is.

Bug: 155497681
Test: existing (new tests forthcoming)
Change-Id: Ifa05a2acd47b6c8c213b44685be31681335d83e6