Improve the checking for live allocation tracking
This CL always show allocation tracking dropdown menu
on supported devices. When the tracking is not ready
during the session startup, the label and menu are
disabled and a tooltip is shown. When the tracking is
ready, the label and menu are enabled and the tooltip
is removed.

For dead sessions, it checkes the device used in the
session instead of the one currently connected to
Studio. A similar issue in native allocation tracking
is fixed also.

Fixes: 150651682
Fixes: 157869432
Fixes: 156670552
Test: updated
Change-Id: I31ce5028b8fab11ea50104cb1e5b0032d5fa1c5e
(cherry picked from commit 24bd1e89952b0dd7ebf61c720cab9a4ba28ba861)