Deprecate TransactionGuard.submitTransaction usage
All the changes here are related to the deprecation of the
TransactionGuard.submitTransaction() methods related to this cherry
picked commit from JetBrains: http://go/androidcommit/53ee762 as part of
the 2020.1 merge.

For detailed information on how to adapt to the deprecated API, and an
explanation of the IDEA threading model (write-safe contexts, locks,
etc) please read this document: http://go/idea-threading-model

1. MotionLayoutAttributesModel.java#createCustomXmlTag()
There is no need to pass a Runnable to submitTransaction() since this
code is already in a write-safe context (since this is triggered by
the dialog). It runs in the EDT, and it uses a write command action to
modify IDE data models.

2. NelePropertiesModel.kt#setPropertyValue()
This code runs in the EDT, and is not in a write-safe context, which is
why the submitTransaction() call is replaced w/ a call to invokeLater().
A Condition<Boolean> is provided to simply wrap the Disposable. Also,
the NlWriteCommandActionUtil already wraps the inner Runnable passed to
it in a write command action.

The following tests have been updated due to the change in
setPropertyValue(), which now requires the EDT events to be pumped
manually using
	PlatformTestUtil.dispatchAllInvocationEventsInIdeEventQueue().

1. ToggleButtonPropertyEditorModelTest.kt
2. DeclaredAttributesInspectorBuilderTest.kt
3. NeleNewPropertyItemTest.kt
4. NelePropertyItemTest.kt
5. NeleFlagsPropertyItemTest.kt

Please note that the following tests were failing before this change
and are still failing with this change.

- MotionLayoutAttributesModelTest.java (5 failures)
- NelePropertiesModelTest.kt (2 failures)

Bug: 156102335

Change-Id: I82e9a053cd40f6132b7a1710465a0c2e1fabdfa9