Customizable selection model
This CL splits the SelectionModel and a DefaultSelectionModel
implementation. This allows DesignSurface users to provide their own
implementations. For now, the Compose preview does not need a selection
model (there is only one NlComponent) so it injects a NopSelectionModel
that disables selection.

The CL also splits the secondary selection into a separate interface so
hopefully we can at some point split it separate from the regular
SelectionModel or to re-work it.

Test: Existing tests passed
Fixes: 158039773
Change-Id: I604cd54d023190b4d1532d81349105c0b22111cb