Properly configure executor and fix NPE
There are two fixes in this CL to fix a test:

1) The executor in ApplicationIdResolver is improperly configured,
such that there is actually no concurrency occuring when
corePoolSize=0 and using an unbounded queue. This forces the test
to time out because only one thread can execute at a time, and
some of the work in the threads of that test deliberately stalls.

The fix is to enable concurrency by increasing the number of core
threads and allow them to time out. This works around the fact
that the default behavior of the executor in use prefers to queue
tasks instead of spawning extra threads (beyond the core threads)
to do work.

2) There is an NPE that may be causing issues.

Fixes: 152354637
Test: Passes DeviceTest without errors.
Change-Id: I019e7da75b89597a9e8ffee4bf24862faaff3cbf