[AUA] allow introspection of refactoring processor
Specifically, external code (such as is found in
AndroidPluginVersionUpdater) may wish to query whether the refactoring
processor has found any instances of code which needs changing.
Decorate each implementation of findUsages() in a subclass of
GradleBuildModelRefactoringProcessor with a call to set a boolean
indicating whether the usages it has found is empty or not.

Bug: 149565531
Test: forthcoming
Change-Id: I7347145239bba23d0de02d0c6dea18a8fca16007