Delayed multi representation initialization
When multiple editors are open at the same time and Studio starts, this
can trigger a large number of renders and increase the load of the
machine. When the previews are not visible, this is unncessary work that
we can avoid until the editor comes into focus.
This CL delays all that initialization and rendering until the user
interacts with that particular editor.

This also enables the representations to handle the case when they
become inactive. As part of a follow-up CL, the Compose preview will
unsubscribe from listeners and/or ignore builds to avoid refreshing when
the user is not seeing the result.

Test: Added new test for the lifecycle
Bug: N/A
Change-Id: I5cbceddab2f30b478b837abf7211defe3a4f42b3