Fix navigation editor background color
After the latest merge the background color for the
navigation editor changed to white. The color was getting
set to white directly in the navigation design surface.
It is unknown why this wasn't taking effect before the merge.
Remove the call to set the color directly and use the default
color scheme that the layout editor uses.

Fixes: 157172421
Test: Manual testing and verification.
Change-Id: If1a007eaf86de5e5b17037a71cf84fd61f6c35e7