[AUA] rename created or configured configurations
A user can configure an existing configuration, including ones that
are obsolete, in the configurations block.  Further, projects with
variants mandate an explicit naming of a configuration (implicitly
creating it in Groovy, explicitly in KotlinScript) if there are
dependencies which should be assigned to configurations with more than
one variant dimension specified.

Find any such explicit naming of configurations under question, and
rename those in the CompileRuntimeConfigurationRefactoringProcessor.

Bug: 149565531
Test: added
Change-Id: Icc3a49492dfcb462201e8d4dc76684c7321bce31