Stabilize CMakeListsTest
CMakeListsTest fails very frequently becuase the
Menu path "Refactor -> Rename..." is not enabled.
As I saw in the screenshot there was some tasks
running in the background and this menu was not enabled.
Wait for background tasks to finish before clicking
the memu path.

Bug: na
Test: Ran this test 300+ times, no failure.
Without this change, it failed many times in 10 times run.

Change-Id: Ie6cb134ed359a72253a803525b9a228c5920403e