[Dsl,.kts] preserve Psi in string-configuration dependencies
In KotlinScript, a dependency is assigned to a non-built-in
configuration by (among other syntaxes) calling a method named by that
configuration as a string.  The parser was too strict in one place
regarding what should be considered as a valid name, leading to the
creation of name elements for those dependencies which had no
association with the underlying Psi, which in turn meant that updating
the name of the configuration (through setConfigurationName() on a
dependency, for example) did nothing.

Delete an additional layer of unquoting of names, which would do no
harm in practice (embedded quotes are among the few characters that
are proscribed in contexts where string names are considered by the
Dsl layer) but is unnecessary in theory.

Change the updating of string names to be the symmetric pair of
"other" names: if we are updating a string name whose parent is the
DependenciesDslElement and whose new name is a known configuration, do
not emit it as a string, but an expression.

The added test case exercises all of these changes.

Bug: 149565531
Test: added
Change-Id: I14f496d020c31d048c3f290852c816bf90a01985