[AUA] fix the upgrade thread handling
The forced upgrade flow is different from the recommended upgrade
flow, and runs on a background thread, which requires us to explicitly
run the refactorings on the EDT.  (AndroidPluginVersionUpdater is
called from a surprising number of places, and even if the forced
upgrade flow is changed it's not clear that none of the other callers
might not also be on background threads.)

Additionally, in a difference between the new code based on
refactoring processors and the code it replaced which operated
directly on the build model, refactoring processors can only operate
in smart mode, and we can be performing this sync in parallel with
indexing.  Run the refactorings, and the subsequent request to resync,
only when smart.

The manual tests involved in checking that this CL passes muster
included:

- running the recommended upgrade flow starting with the popup balloon
  notification after a successful sync;
- running the recommended upgrade flow starting with the AndroidLint
  intention;
- running the forced upgrade flow from a sync started after indexing
  was complete;
- running the forced upgrade flow from a sync started during indexing.

Bug: 159840300
Test: manual
Change-Id: If47bbcc15e33edf8084570908c29794bb69d8a2b