Adds Android Studio run configuration
This allows Android Studio to be run using the prebuilt
IntelliJ SDK.

It uses idea.home to be where the prebuilt is, and we set
the path selector to AndroidStudioDev, this way the settings
are not .gitignored with the source code but alongside
the real settings.

Note this allows Android Studio to "open" but many things are
still broken as they are looked for relative to the home path.

All the IJ binaries are found correctly (such as the fsnotifier).

Test: N/A
Bug: 156497102
Change-Id: Iee9c10f2ef955c868110d9b0b5fce3d6b86291d9