[AUA] don't update configurations in buildscript
The buildscript dependencies are those needed for executing the rest
of the build, not those related to the project.  Out-of-the-box, the
only configuration available at that point is classpath, so there
should be no obsolete configurations there anyway; even if we do find
one that is named `compile` or `runtime`, there is no guarantee as far
as I can see that that configuration would have the same identity as
the deprecated ones.  (It could come from some other build-time plugin
with its own semantics).

Bug: n/a
Test: added
Change-Id: Ie3997a513959568b60164f9de2f0d8d1c5d823d2