Add support for Compose package migration
The CL adds support for both Compose namespaces in preparation for the
library migration.
The code added will be here temporarily and can probably removed after
alpha. We just need to ensure that the code support the transition
period where both names exist in the wild.

This CL does not cover the tests using AndroidTestCase. They need to be
migrated to use Rules to support the migration and that will be done as
a separate CL.

Test: Updated tests with support for both namespaces.
Bug: 160404330
Change-Id: Ia81f09e612d3fad629d6a14842b055911216c04e