Make the view not move around
When the depth of the 3d view changes (that is, more or fewer layers are
shown) the size of the canvas changes. Since the views are always drawn
centered in the canvas, this means that if the scroll position isn't
adjusted when the canvas size changes, the view moves around in the
window.
Adjusting the scroll position to try to keep the root view in the same
place when the view is updated.

Test: added a test
Fixes: 155920880
Change-Id: I000ba535070825137249e8ea8ac5e22a230f0d04
(cherry picked from commit ba0a551ce6e24ea99c14e32e7a90c6f5d897b565)