[AUA] fabric to crashlytics refactor component
This upgrade assistant component only handles the Gradle build parts
of the move from Fabric to Firebase Crashlytics.  The need for a
google-services json file, and the need to update code, will be
handled by some combination of documentation (pointed at by the
upgrade component in the dialog box), the Firebase Crashlytics SDK,
and the Studio What's New assistant.

This change only uses the refactoring processor within the flow of an
AGP upgrade, but as with the compile/runtime deprecated
configurations, it could also be activated from a lint check.

Bug: 156100314
Test: added
Change-Id: I30829a46ffe92abac2250f119c39551563bb7c80