Enable test by not using spy project
Using a spy project causes an exception in the tests when calling
getComponent for ModuleManager after IDEA 193.2956.37, this is because
the real project is not the same object as the spied project. The reason
to use a spy project is to simulate it was not initialized.

This CL removes the usage of the spy project by adding a test flag to
assume the project is not initialized. There are also some minor changes
to remove unused code and deprecated calls.

Test: Existing test code re enabled
Bug: 144931471
Change-Id: I7511234a66c25e15950d73d33c6e479c7e2b6a53