Auto-dispose usage trackers
This CL forces the DesignerUsageTrackerManager key to be a Disposable.
This ensures that if the key is disposed, then, the usage tracker will
be disposed immediately.
This will heap speeding up the release of the trackers. The will improve
the current case where the trackers sometimes link to the key which
prevents the cache from being released.

Fixes: 159355159
Test: Updated the test.
Change-Id: Ic4e938982ea36c6bc938b197cce837ea28a4ffb0