Remove FileHandler extension point
This extension point was originally added to download
additional files from Device Explorer, other than the one
selected by the user.

Some SQLite database use write ahead log, in order to open them
multiple files are needed: the .db file, the shm and wal files.

FileHandler was used to download shm and wal files
when the user initiate the download of the .db file.

This not necessary anymore, since DatabaseInspector
has the ability to initiate the download of files,
therefore can explicitly download the additional files.

Test: NA
Bug: 160868563
Change-Id: I8bdf519ece08e97aaaad347e6dc5f8b117c070d9