[AUA] group usages in preview by component processor
This unconditionally creates one usage group per processor in the
preview window, at the outermost level (nearest the root of the tree
of usages).  It is only active for the Upgrade Assistant and related
refactorings because it will only return non-null if the usage has a
usageInfo which is a GradleBuildModelUsageInfo; it will return null
for all other usages.

Bug: 149565531
Test: manual, added
Change-Id: If8d4f0cd137df62197358c048e7efb4e8a2be95a