Configure the right module when importing UTP results
This commit tries to find the module with the correct package name when
importing UTP results. It would enable the double click to jump to line
feature in test result window.

Also, we need to have some minor refacturing over UtpTestResultAdapter
because now we need to parse the protobuf before constructing
AndroidTestResultListener.

Bug: 161565889
Test: ImportUtpResultActionTest, UtpTestResultAdapterTest
Change-Id: I7021445cc86e889d3a230453979fba5939e41997